requirements.txt
discord.py
steam
selenium
urllib3==1.24